refactor: [move] application/service 안으로 이동, CommentWriterService userRepository, ReviewRepository (인터페이스에) 의존하도록 수정
