+++ src/main/java/site/travellaboratory/be/comment/presentation/controller/CommentLikeController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import site.travellaboratory.be.comment.application.CommentLikeService;
+import site.travellaboratory.be.comment.application.service.CommentLikeService;
 import site.travellaboratory.be.common.annotation.UserId;
 import site.travellaboratory.be.comment.domain.enums.CommentLikeStatus;
 import site.travellaboratory.be.comment.presentation.response.like.CommentToggleLikeResponse;
