+++ src/main/java/site/travellaboratory/be/comment/presentation/controller/CommentWriterController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import site.travellaboratory.be.comment.application.CommentWriterService;
+import site.travellaboratory.be.comment.application.service.CommentWriterService;
 import site.travellaboratory.be.comment.presentation.response.writer.CommentDeleteResponse;
 import site.travellaboratory.be.comment.domain.request.CommentSaveRequest;
 import site.travellaboratory.be.common.annotation.UserId;
@@ -30,7 +30,7 @@ public class CommentWriterController {
         @UserId Long userId,
         @Valid @RequestBody CommentSaveRequest commentSaveRequest
     ) {
-        Long result = commentWriterService.saveComment(userId, commentSaveRequest);
+        Long result = commentWriterService.save(userId, commentSaveRequest);
         return new ResponseEntity<>(CommentSaveResponse.from(result), HttpStatus.CREATED);
     }
 
@@ -40,7 +40,7 @@ public class CommentWriterController {
         @PathVariable(name = "commentId") Long commentId,
         @Valid @RequestBody CommentUpdateRequest commentUpdateRequest
     ) {
-        Long result = commentWriterService.updateComment(userId, commentId, commentUpdateRequest);
+        Long result = commentWriterService.update(userId, commentId, commentUpdateRequest);
         return ResponseEntity.ok(CommentUpdateResponse.from(result));
     }
 
@@ -49,7 +49,7 @@ public class CommentWriterController {
         @UserId Long userId,
         @PathVariable(name = "commentId") Long commentId
     ) {
-        boolean result = commentWriterService.deleteComment(userId, commentId);
+        boolean result = commentWriterService.delete(userId, commentId);
         return ResponseEntity.ok(CommentDeleteResponse.from(result));
     }
 }
